added substep functionality
add class 'substep' to any element in a 'step'. Like the normal steps
substeps will be in on of the three temporal states 'future', 'present'
or 'past'.
If a data-group is specified all substeps in this group are shown at the
same time.
data-resetsubsteps="false" let all substeps past when you reenter a step